Pcvl 602 add tomography pdisplay_to_file (#323)
* add save tomography method & add UT
