+++ tests/_test_utils.py
@@ -27,9 +27,19 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from perceval.utils import StateVector, SVDistribution
+import re
+from typing import Type
+from pathlib import Path
 
 import pytest
+
+import perceval as pcvl
+from perceval.utils import StateVector, SVDistribution
+from perceval.rendering import Format
+from perceval.rendering.circuit import ASkin, PhysSkin
+from perceval.algorithm import ProcessTomography
+
+TEST_IMG_DIR = Path(__file__).resolve().parent / 'imgs'
 
 
 def strip_line_12(s: str) -> str:
@@ -78,6 +88,78 @@ def assert_svd_close(lhsvd, rhsvd):
 
 def  dict2svd(d: dict):
     return SVDistribution({StateVector(k): v for k, v in d.items()})
+
+
+@pytest.fixture(scope="session")
+def save_figs(pytestconfig):
+    return pytestconfig.getoption("save_figs")
+
+
+def _norm(svg):
+    svg = svg.replace(" \n", "\n")
+    svg = re.sub(r'url\(#.*?\)', 'url(#staticClipPath)', svg)
+    svg = re.sub(r'<clipPath id=".*?">', '<clipPath id="staticClipPath">', svg)
+    svg = re.sub(r'<dc:date>(.*)</dc:date>', '<dc:date></dc:date>', svg)
+    svg = re.sub(r'<dc:title>(.*)</dc:title>', '<dc:title></dc:title>', svg)
+    return svg
+
+
+def _check_svg(test_path, ref_path, collection: str):
+    with open(test_path) as f_test:
+        test = _norm("".join(f_test.readlines()))
+    with open(ref_path) as f_ref:
+        ref = _norm("".join(f_ref.readlines()))
+    m_test = re.search(rf'<g id="{collection}Collection.*?>((.|\n)*?)</g>', test)
+    m_ref = re.search(rf'<g id="{collection}Collection.*?>((.|\n)*?)</g>', ref)
+    if not m_test:
+        return False, "cannot find patch in test"
+    if not m_ref:
+        return False, "cannot find patch in ref"
+    m_test_patch = re.sub(r'url\(#.*?\)', "url()", m_test.group(1))
+    m_ref_patch = re.sub(r'url\(#.*?\)', "url()", m_ref.group(1))
+    if m_test_patch != m_ref_patch:
+        return False, "test and ref are different"
+    return True, "ok"
+
+
+def _check_circuit(test_path, ref_path):
+    return _check_svg(test_path, ref_path, 'Patch')
+
+
+def _check_qpt(test_path, ref_path):
+    return _check_svg(test_path, ref_path, 'Line3D')
+
+
+def _save_or_check(c, tmp_path, circuit_name, save_figs, recursive=False, compact=False,
+                   skin_type: Type[ASkin] = PhysSkin) -> None:
+    img_path = (TEST_IMG_DIR if save_figs else tmp_path) / \
+        Path(circuit_name + ".svg")
+    skin = skin_type(compact)
+
+    if isinstance(c, ProcessTomography):
+        pcvl.pdisplay_to_file(c, img_path, output_format=Format.MPLOT)
+    elif isinstance(c, pcvl.AComponent) or isinstance(c, pcvl.components.ACircuit) or isinstance(c, pcvl.AProcessor):
+        pcvl.pdisplay_to_file(c, img_path, output_format=Format.MPLOT,
+                              recursive=recursive, skin=skin)
+    else:
+        raise NotImplementedError(f"_save_or_check not implemented for {type(c)}")
+
+    if save_figs:
+        with open(img_path) as f_saved:
+            saved = "".join(f_saved.readlines())
+        saved = _norm(saved)
+        with open(img_path, "w") as fw_saved:
+            fw_saved.write(saved)
+    else:
+        if isinstance(c, ProcessTomography):
+            ok, msg = _check_qpt(img_path, TEST_IMG_DIR /
+                                 Path(circuit_name + ".svg"))
+        elif isinstance(c, pcvl.AComponent) or isinstance(c, pcvl.components.ACircuit) or isinstance(c, pcvl.AProcessor):
+            ok, msg = _check_circuit(img_path, TEST_IMG_DIR /
+                                     Path(circuit_name + ".svg"))
+        else:
+            raise NotImplementedError(f"_save_or_check not implemented for {type(c)}")
+        assert ok, msg
 
 
 if __name__ == "__main__":
