+++ tests/test_tomography.py
@@ -27,6 +27,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+import sys
+from pathlib import Path
+
 import pytest
 import numpy as np
 from scipy.stats import unitary_group
@@ -39,8 +42,10 @@ from perceval.algorithm import ProcessTomography, StateTomography
 from perceval.algorithm.tomography.tomography_utils import is_physical, get_preparation_circuit, \
     _generate_pauli_index, _vector_to_sq_matrix, _matrix_to_vector, _matrix_basis, _coef_linear_decomp
 
+from _test_utils import save_figs, _save_or_check
 
 CNOT_TARGET = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]], dtype='complex_')
+TEST_IMG_DIR = Path(__file__).resolve().parent / 'imgs'
 
 
 @pytest.mark.parametrize("pauli_gate", [PauliType.I, PauliType.X, PauliType.Y, PauliType.Z])
@@ -104,7 +109,7 @@ def test_fidelity_random_op():
     assert random_op_fidelity == pytest.approx(1)
 
 
-def test_chi_cnot_is_physical():
+def test_chi_cnot_is_physical_and_display(tmp_path, save_figs):
     cnot_p = catalog["klm cnot"].build_processor()
 
     qpt = ProcessTomography(operator_processor=cnot_p)
@@ -116,6 +121,8 @@ def test_chi_cnot_is_physical():
     assert res['Hermitian'] is True  # if Chi is Hermitian
     assert res['Completely Positive'] is True  # if input Chi is Completely Positive
 
+    # display
+    _save_or_check(qpt, tmp_path, sys._getframe().f_code.co_name, save_figs)
 
 def test_processor_odd_modes():
     # tests that a generic processor with odd number of modes does not work
